게임 조회 랭킹 컨트롤러에서 게임 컨트롤러로 변경
