+++ src/main/java/com/example/jsgamesbackendmain/Controller/GameController.java
@@ -30,6 +30,16 @@ public class GameController {
         return gameService.postGame(gameCreateRequestDTO);
     }
 
+    // 게임 목록 조회 API
+    @Operation(summary = "게임 목록 조회 (페이징) ", description =
+            "# 게임 목록 조회  \n" +
+                    "## 조회수 순으로 내림차순 정렬됩니다."
+    )
+    @GetMapping("/games")
+    public List<GameListResponseDTO> listGames(@Parameter Integer page, @Parameter Integer size) {
+        return gameService.listGames(page, size);
+    }
+
     // 특정 게임 조회 API
     @Operation(summary = "GameId로 게임 조회", description =
         "# GameId로 게임 조회시 조회수가 1 증가합니다."
